requirements.txt
workcell
tweety-ns